Level name consistency update
