--- Assets/AddressableAssetsData/link.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: adb011f647b5e7f4a8fd2763aca24a98
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
