--- Assets/Scenes/Apples01.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b7ceedfa10efa0c4ebb25e80930620ba
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
