--- Assets/Scripts/Gameplay/Apples.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c6a3252f5764ff24f8bff2564911ac95
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
