--- Assets/Scripts/Gameplay/Apples/Prize.cs
@@ -1,42 +1,0 @@
-using UnityEngine;
-using Mirror;
-
-namespace MirrorBasics
-{
-    public class Prize : NetworkBehaviour
-    {
-        public bool available = true;
-
-        [SerializeField]  public string type;
-
-        public override void OnStartServer ()
-        {
-            
-        }
-
-// Server check if player can pick up reward
-    [ServerCallback]
-        void OnTriggerEnter(Collider other)
-        {
-            if (other.gameObject.CompareTag("Player"))
-            {
-                if (other.gameObject.GetComponent<PlayerScore>().hasItem == true) {return;}
-                else
-                PickUpItem(other.gameObject);               
-            }
-        }
-
-// Server adds reward to a player and destroy pickup
-    [ServerCallback]
-        public void PickUpItem(GameObject player)
-        {
-            if (type == "Reward")
-            player.GetComponent<PlayerScore>().hasItem = true;
-            NetworkServer.Destroy(gameObject);
-
-            if (type =="Stealing")
-            player.GetComponent<PlayerScore>().canSteal = true;
-            NetworkServer.Destroy(gameObject);
-        }
-    }
-}
