--- Assets/Scripts/Gameplay/Apples/Prize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 280c8879ff8849743a0fd55f4d9b269e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
