+++ Assets/Scripts/Gameplay/Farmapples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c6a3252f5764ff24f8bff2564911ac95
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
