+++ Assets/Scripts/Gameplay/Farmapples/TeamBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e179d45c5a98d4740bc37691123f662a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
